Fun with markov chains
Contains a now quite heavily modified version of Chris Hawkins's Markov function

The algorithm is pretty much the same, but I've made it a bit easier to integrate
and now you can pass a names file to it and let it make neat chains off of that.

This also contains two 'corpus' files for markov chain generation.

They honestly might be too big and general; I was being hasty. Could definitely
be fine-tuned.

They are:
A file with the names of every town in Massachusets, New Hampshire, and Connecticut
and a file with the entire old english Beowulf poem, less any word with four or
fewer letters.

Did all of the transformation on Beowulf with vim, google, and a steady hand.