'details' button on ship crafting keyed into codex
A lot needed to happen for this. One thing is making sure that switching to the codex won't
mess up the pause state. Another turns out to be moving the notion of paths out of the
codex UI class and handling it (mostly) in Data. Also ships need to know where in the
codex they fit so they need an inherant faction (might later replace this with, say,
a manufacturer or some such.)